wip: setup screen.
