--- js/background/functions.js
@@ -1,24 +1,0 @@
-'use strict';
-
-export function bla() {
-//     var storage = new API.Storage();
-//     var _self = this;
-//     var _window = {};
-
-// 		let url = API.runtime.getURL("/html/browser_action/browser_action.html");//extension + '://' + api.runtime.id + '/html/browser_action/browser_action.html';
-		console.log("urrr;");
-// 		console.log(url);
-
-//     // API.runtime.onConnect.addListener(function (port) {
-
-//     //     port.onMessage.addListener(function (msg) {
-//     //         if (msg === 'credential_amount') {
-//     //             port.postMessage('credential_amount:' + local_credentials.length);
-//     //         }
-
-//     //     });
-
-//     // });
-}
-
-// console.log("BUMS!");
