+++ js/background/service-worker.js
@@ -1,8 +1,10 @@
 'use strict';
 
 import * as browserapi from '../lib/API/base.js';
-// import * as papi from '../lib/api.js';
-import * as func from './functions.js';
+import * as setup from './functions/setup.js';
+import * as settings from './functions/settings.js';
+
+let importedFunctions = [setup, settings];
 
 let API = browserapi.initAPI();
 // let PAPI = papi.initPAPI();
@@ -17,22 +19,29 @@ API.api.runtime.onInstalled.addListener(function () {
 
 });
 
-API.api.runtime.onMessage.addListener((message, sender, sendResponse) => {
+API.api.runtime.onMessage.addListener(async (message, sender, sendResponse) => {
 	if (!message || !message.hasOwnProperty('method')) {
 		return;
 	}
 
 	var result = false;
 
-	if (typeof func[message.method] === "function") {
-		result = func[message.method](message.args);
+	// Anstatt alle background/functions/bla.js importieren zu müssen, will ich hier eine "caller funktuon" in der main.js aufrufen, welche das überminnt.
+	for (const element of importedFunctions) {
+		if (typeof element[message.method] === "function") {
+			if (element[message.method].constructor.name === "AsyncFunction") {
+				result = await element[message.method](message.args);
+			} else {
+				result = element[message.method](message.args);
+			}
+			break;
+		}
+	}
+
+	if (result !== false) {
+		return Promise.resolve(result);
 	} else {
 		console.warn('[NOT FOUND] Method call', message.method, 'args: ', message.args);
 	}
-	sendResponse(result);
 
-	// TODO
-	// if (msg === 'credential_amount') {
-	// port.postMessage('credential_amount:' + local_credentials.length);
-	// }
 });
