+++ js/background/service/background.js
@@ -147,31 +147,31 @@ var background = (function () {
 
     _self.getSetting = getSetting;
 
-    function saveSettings(settings, cb) {
-        for (var i = 0; i < encryptedFieldSettings.length; i++) {
-            var field = encryptedFieldSettings[i];
-            settings[field] = PAPI.encryptString(JSON.stringify(settings[field]), master_password);
-        }
+    // function saveSettings(settings, cb) {
+    //     for (var i = 0; i < encryptedFieldSettings.length; i++) {
+    //         var field = encryptedFieldSettings[i];
+    //         settings[field] = PAPI.encryptString(JSON.stringify(settings[field]), master_password);
+    //     }
 
-        if (!settings.hasOwnProperty('ignored_sites')) {
-            settings.ignored_sites = [];
-        }
+    //     if (!settings.hasOwnProperty('ignored_sites')) {
+    //         settings.ignored_sites = [];
+    //     }
 
-        if (!_self.settings.hasOwnProperty('password_picker_first_tab')) {
-            _self.settings.disable_browser_autofill = 'list';
-        }
+    //     if (!_self.settings.hasOwnProperty('password_picker_first_tab')) {
+    //         _self.settings.disable_browser_autofill = 'list';
+    //     }
 
-        //window.settings contains the run-time settings
-        _self.settings = settings;
+    //     //window.settings contains the run-time settings
+    //     _self.settings = settings;
 
 
-        storage.set('settings', settings).then(function () {
-            getSettings();
-        });
+    //     storage.set('settings', settings).then(function () {
+    //         getSettings();
+    //     });
 
-    }
+    // }
 
-    _self.saveSettings = saveSettings;
+    // _self.saveSettings = saveSettings;
 
     // function resetSettings() {
     //     storage.set('settings', {});
